pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Kyle1297 <kyle.griffin.1297@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-restql = "^0.11.0"
django-cors-headers = "^3.7.0"
Django = "^3.1.6"
python-dotenv = "^0.15.0"
boto3 = "^1.17.3"
django-storages = "^1.11.1"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.0.4"
djangorestframework = "^3.12.2"
django-filter = "^2.4.0"
django-rest-knox = "^4.1.0"
requests = "^2.25.1"
channels = "^3.0.3"
redis = "^3.5.3"
daphne = "^3.0.1"
channels-redis = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
